Change X-MEDIA-SEQUENCE type from int to long

+++ src/main/java/io/lindstrom/m3u8/parser/MediaPlaylistParser.java
@@ -94,7 +94,7 @@ public class MediaPlaylistParser extends AbstractPlaylistParser<MediaPlaylist, M
                 break;
 
             case EXT_X_MEDIA_SEQUENCE:
-                builder.mediaSequence(Integer.parseInt(attributes));
+                builder.mediaSequence(Long.parseLong(attributes));
                 break;
 
             case EXT_X_KEY:
